pyproject.toml
[build-system]
build-backend = "pdm.backend"
requires = ["pdm-backend"]

[project]
name = "arxiv-explorer"
version = "0.1.0"
description = "Explore arXiv papers with semantic search powered by polars-fastembed"
readme = "README.md"
authors = [
    { name = "Louis Maddox", email = "louismmx@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.5",
    "polars-fastembed[cuda]>=0.1.10",
    "umap-learn>=0.5.9.post2",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "tqdm>=4.67.1",
    "huggingface-hub>=1.2.3",
]

[project.scripts]
arxiv-partition = "arxiv_explorer.partition:main"
arxiv-embed = "arxiv_explorer:run_embed"
arxiv-serve = "arxiv_explorer:run_serve"

[dependency-groups]
dev = [
    "ty>=0.0.3",
]
partition = [
    "datasets>=4.4.2",
]

[tool.uv]
default-groups = ["dev", "partition"]

